+++ endpoint/thread/CMakeLists.txt
@@ -22,3 +22,7 @@ if(PRIVMX_WERROR)
     target_compile_options(privmxendpointthread PRIVATE -Werror)
 endif()
 set_target_properties(privmxendpointthread PROPERTIES COMPILE_DEFINITIONS "MINIMAL_BUILD")
+
+if(PRIVMX_BUILD_ENDPOINT_INTERFACE)
+    add_subdirectory(cinterface)
+endif(PRIVMX_BUILD_ENDPOINT_INTERFACE)
